Timeline: Use some better number types
The scale parameters are converted to single precision float anyway as
soon as we do something with them. We might as well change all the
methods to be single precision, too.

The min/max row values should really be 64bit as 32bit values are not
enough to express memory usage.

+++ src/libs/timeline/timelineitemsrenderpass.h
@@ -95,7 +95,7 @@ public:
     static const TimelineItemsRenderPass *instance();
     State *update(const TimelineAbstractRenderer *renderer, const TimelineRenderState *parentState,
                   State *state, int firstIndex, int lastIndex, bool stateChanged,
-                  qreal spacing) const;
+                  float spacing) const;
 protected:
     TimelineItemsRenderPass();
 };
