+++ src/libs/timeline/timelineselectionrenderpass.h
@@ -38,7 +38,7 @@ public:
 
     State *update(const TimelineAbstractRenderer *renderer, const TimelineRenderState *parentState,
                   State *state, int firstIndex, int lastIndex, bool stateChanged,
-                  qreal spacing) const;
+                  float spacing) const;
 
 protected:
     TimelineSelectionRenderPass();
